prime now works with Int instead of Double, makes more sense

+++ e10.hs
@@ -3,5 +3,5 @@ import Utils
 -- sum all primes with a value < x
 -- use 2000000 for problem #10
 
-primeSum :: Double -> Double
+primeSum :: Int -> Int
 primeSum x = sum $ takeWhile ((>) x) $ filter prime [2..]
